% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Deriv_cpp}
\alias{Deriv_cpp}
\title{Symbolic Differentiation (Supports Vectors of Variables)}
\usage{
Deriv_cpp(expr, x, env = parent.frame(), drule = Deriv::drule)
}
\arguments{
\item{expr}{An R language object, symbol, or numeric constant to differentiate.}

\item{x}{A character vector, symbol, or named vector of targets.}

\item{env}{An environment to resolve function calls in (defaults to parent.frame()).}

\item{drule}{An environment containing derivative rules (cf. Details in \code{\link{Deriv}} for syntax rules).}
}
\value{
A simplified R expression (or a named c() call for multiple inputs).
}
\description{
Computes the symbolic derivative of an R expression with respect to one
or more target variables. If a single target is provided, it returns a 
scalar expression. If multiple targets are provided, it returns a named c() call.
}
\details{
This version is much faster for simple expressions but has less features than its R counterpart Deriv().
}
\seealso{
\code{\link{Deriv}}
}
