% Generated by roxygen2 (4.0.2): do not edit by hand
\name{format1}
\alias{format1}
\title{Wrapper for base::format() function}
\usage{
format1(expr)
}
\arguments{
\item{expr}{An expression or symbol or language to be converted to a string.}
}
\value{
A character vector of length 1 contrary to base::format() which
 can split its output over several lines.
}
\description{
Wrapper for base::format() function
}

