% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Simplify_cpp}
\alias{Simplify_cpp}
\title{Simplify Symbolic Expressions}
\usage{
Simplify_cpp(expr)
}
\arguments{
\item{expr}{An R language object, symbol, or numeric constant to simplify.}
}
\value{
A simplified R expression or scalar numeric value.
}
\description{
Simplify Symbolic Expressions
}
